et ieta attempt 12

+++ makeet_ieta_plot.cxx
@@ -52,7 +52,7 @@ void makeet_ieta_plot(std::string run, double cut_et, int cut_ieta)
     
     std::vector<double> eta_values = {0.,0.087,0.174,0.261,0.348,0.435,0.522,0.609,0.695,0.783,0.87,0.957,1.044,1.131,1.218,1.305,1.392,1.479,1.566,1.653,1.74,1.83,1.93,2.043,2.172,2.322,2.5,2.65,3.0};
     
-    std::vector<int> etbins = {5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80};
+    std::vector<double> etbins = {5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80};
 
     
     std::vector<std::string> inDir;
